[ feat ] - added easteregg
