+++ src/components/NavBar.js
@@ -1,4 +1,4 @@
-import { useState, useEffect, useContext } from "react";
+import { useState, useContext } from "react";
 import { Navbar, Nav, Container } from "react-bootstrap";
 import lego from "../assets/icons/lego.svg";
 import lego_white from "../assets/icons/lego-white.svg";
@@ -15,27 +15,13 @@ import { BrowserRouter as Router } from "react-router-dom";
 
 export const NavBar = () => {
   const [activeLink, setActiveLink] = useState("home");
-  const [scrolled, setScrolled] = useState(false);
-
-  useEffect(() => {
-    const onScroll = () => {
-      if (window.scrollY > 50) {
-        setScrolled(true);
-      } else {
-        setScrolled(false);
-      }
-    };
-
-    window.addEventListener("scroll", onScroll);
-
-    return () => window.removeEventListener("scroll", onScroll);
-  }, []);
 
   const onUpdateActiveLink = (value) => {
     setActiveLink(value);
   };
 
   const theme = useContext(AppContext).theme;
+  const scrolled = useContext(AppContext).scrolled;
   const toggleNav = useContext(AppContext).toggleNav;
   const setDarkMode = useContext(AppContext).setDarkMode;
   const setLightMode = useContext(AppContext).setLightMode;
